requirements.txt
crccheck==1.3.0
pyserial==3.5
tqdm==4.64.0
